Refactor cli command
