+++ controller/command/cmd.go
@@ -1,55 +1,16 @@
-package main
+package command
 
 import (
 	xkube "github.com/crossplane-contrib/provider-kubernetes/apis/object/v1alpha1"
 	"github.com/go-logr/logr"
 	"github.com/urfave/cli/v2"
-	appcatv1 "github.com/vshn/appcat-apiserver/apis/appcat/v1"
-	"github.com/vshn/appcat-apiserver/apiserver/appcat"
-	"github.com/vshn/appcat-apiserver/apiserver/vshn/postgres"
 	"github.com/vshn/appcat-apiserver/controller"
 	vshnv1 "github.com/vshn/component-appcat/apis/vshn/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/klog"
-	"sigs.k8s.io/apiserver-runtime/pkg/builder"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 )
-
-func runApiServer(ctx *cli.Context) error {
-	appcatEnabled := ctx.Bool("appcat-handler")
-	vshnBackupsEnabled := ctx.Bool("postgres-backup-handler")
-
-	if !appcatEnabled && !vshnBackupsEnabled {
-		klog.Fatal("Handlers are not enabled, please set at least one of APPCAT_HANDLER_ENABLED | VSHN_POSTGRES_BACKUP_HANDLER_ENABLED env variables to true")
-	}
-
-	b := builder.APIServer
-
-	if appcatEnabled {
-		b.WithResourceAndHandler(&appcatv1.AppCat{}, appcat.New())
-	}
-
-	if vshnBackupsEnabled {
-		b.WithResourceAndHandler(&appcatv1.VSHNPostgresBackup{}, postgres.New())
-	}
-
-	b.WithoutEtcd().
-		ExposeLoopbackAuthorizer().
-		ExposeLoopbackMasterClientConfig()
-
-	cmd, err := b.Build()
-	if err != nil {
-		klog.Fatal(err)
-	}
-
-	if err := cmd.Execute(); err != nil {
-		klog.Fatal(err)
-	}
-
-	return nil
-}
 
 var s = runtime.NewScheme()
 
@@ -59,8 +20,34 @@ func init() {
 	_ = vshnv1.SchemeBuilder.SchemeBuilder.AddToScheme(s)
 }
 
+func Command() *cli.Command {
+	return &cli.Command{
+		Name:   "controller",
+		Usage:  "A controller to manage PostgreSQL instance deletion",
+		Action: execute,
+		Flags: []cli.Flag{
+			&cli.StringFlag{
+				Name:  "metrics-addr",
+				Value: ":8080",
+				Usage: "The address the metric endpoint binds to.",
+			},
+			&cli.StringFlag{
+				Name:  "health-addr",
+				Value: ":8081",
+				Usage: "The address the probe endpoint binds to.",
+			},
+			&cli.BoolFlag{
+				Name:  "leader-elect",
+				Value: false,
+				Usage: "Enable leader election for controller manager. " +
+					"Enabling this will ensure there is only one active controller manager.",
+			},
+		},
+	}
+}
+
 // Run will run the controller mode of the composition function runner.
-func runController(cli *cli.Context) error {
+func execute(cli *cli.Context) error {
 
 	log := logr.FromContextOrDiscard(cli.Context)
